pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bedrock-copilot"
version = "3.0.0"
description = "AWS-powered multi-agent code assistant with enterprise security"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Bedrock Copilot Contributors", email = "dev@bedrockcopilot.com"}
]
keywords = ["aws", "bedrock", "copilot", "code-assistant", "ai", "multi-agent"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Tools",
]
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.26.0",
    "botocore>=1.29.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "chromadb>=0.4.0",
    "filelock>=3.12.0",
    "docker>=6.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "bidict>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.4.0",
    "ruff>=0.1.0",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.2.0",
    "myst-parser>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/bedrock-copilot"
Documentation = "https://bedrock-copilot.readthedocs.io"
Repository = "https://github.com/yourusername/bedrock-copilot"
Issues = "https://github.com/yourusername/bedrock-copilot/issues"

[project.scripts]
copilot = "src.main:main"

[tool.setuptools]
package-dir = {"" = "."}
packages = ["src", "src.config", "src.core", "src.core.security", "src.core.features", "src.core.storage", "src.ui"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_mode = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [".git", "__pycache__", "build", "dist"]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing -v"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
